enhancement: CompletionExecutor(Chat Completion 실행 모듈) 모듈화
