+++ app/util/__init__.py
@@ -1,0 +1,1 @@
+from app.util.completion_executor import CompletionExecutor
